+++ backend/src/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod auth;
+pub mod email;
+
+pub use auth::{login, logout, protected, public, refresh, register};
+pub use email::{resend_verification, verify_email};
