+++ backend/src/models/mod.rs
@@ -1,0 +1,5 @@
+pub mod token;
+pub mod user;
+
+pub use token::{MyClaims, RefreshToken};
+pub use user::{Role, User, UserBasic};
